Add WTC Auth

+++ app/Http/Controllers/CourseController.php
@@ -21,6 +21,12 @@ class CourseController extends Controller
      */
     public function index(Request $request)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         $courses = Course::orderBy('updated_at', 'DESC')
             ->orderBy('created_at', 'DESC')
             ->paginate(5);
@@ -38,6 +44,12 @@ class CourseController extends Controller
      */
     public function search(Request $request)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         if ($request->query('query') == '') {
             $courses = Course::orderBy('updated_at', 'DESC')->paginate(5);
             if ($request->ajax()) {
@@ -88,6 +100,12 @@ class CourseController extends Controller
      */
     public function create()
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         return view('courses.create');
     }
 
@@ -99,6 +117,12 @@ class CourseController extends Controller
      */
     public function store(Request $request)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         $rules = [
             'module_no' => 'required',
             'module_title' => 'required',
@@ -129,6 +153,12 @@ class CourseController extends Controller
      */
     public function show(Course $course)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         // decrypt uni_identifiers and grades
         $course->gradings->each(function ($grading) {
             $grading->decryptUniIdentifier();
@@ -150,6 +180,12 @@ class CourseController extends Controller
      */
     public function edit(Course $course)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         return view('courses.edit', compact('course'));
     }
 
@@ -162,6 +198,12 @@ class CourseController extends Controller
      */
     public function update(Request $request, Course $course)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         $rules = [
             'module_no' => 'required',
             'module_title' => 'required',
@@ -260,6 +302,12 @@ class CourseController extends Controller
      */
     public function destroy(Request $request, Course $course)
     {
+        if (!$this->isAuthenticated()) {
+            return view('login');
+        }
+        if (!$this->authorize('mitarbeiter')) {
+            abort(403);
+        }
         $course->delete();
         if ($request->ajax()) {
             return $course;
